requirements.txt
tzlocal~=2.0.0
robotframework==3.1.2
six
zeep==3.1.0
dictdiffer==0.8.0
RESTinstance==1.0.2
openpyxl
xmltodict==0.12.0
robotframework-mongodb-library3
robotframework-jsonlibrary==0.3.1
robotframework-soaplibrary==0.4
robotframework-requests==0.6.2
robotframework-databaselibrary==1.2.4
robotframework-seleniumlibrary==4.5.0
requests~=2.23.0
pytz~=2019.3
pprintpp==0.4.0
pyodbc==4.0.30
selenium==3.141.0
PyMySQL==0.10.1